+++ src/Exception/InvalidSettingException.php
@@ -5,6 +5,6 @@ namespace Drupal\os2forms_organisation\Exception;
 /**
  * Invalid setting exception.
  */
-class InvalidSettingException extends \Exception {
+class InvalidSettingException extends Exception {
 
 }