+++ src/Helper/OrganisationHelper.php
@@ -2,22 +2,19 @@
 
 namespace Drupal\os2forms_organisation\Helper;
 
+use ItkDev\Serviceplatformen\Service\SF1500\AbstractService;
+use ItkDev\Serviceplatformen\Service\SF1500\AdresseService;
+use ItkDev\Serviceplatformen\Service\SF1500\BrugerService;
+use ItkDev\Serviceplatformen\Service\SF1500\Model\Bruger;
+use ItkDev\Serviceplatformen\Service\SF1500\PersonService;
 use ItkDev\Serviceplatformen\Service\SF1500\SF1500;
-use ItkDev\Serviceplatformen\Service\SF1500\SF1500XMLBuilder;
 use ItkDev\Serviceplatformen\Service\SF1514\SF1514;
 use ItkDev\Serviceplatformen\Service\SoapClient;
-use Symfony\Component\PropertyAccess\PropertyAccessor;
 
 /**
  * Organisation Helper service.
  */
 class OrganisationHelper {
-  /**
-   * The PropertyAccessor.
-   *
-   * @var \Symfony\Component\PropertyAccess\PropertyAccessor
-   */
-  private PropertyAccessor $propertyAccessor;
 
   /**
    * The Settings.
@@ -43,8 +40,7 @@ class OrganisationHelper {
   /**
    * Constructor.
    */
-  public function __construct(PropertyAccessor $propertyAccessor, CertificateLocatorHelper $certificateLocator, Settings $settings) {
-    $this->propertyAccessor = $propertyAccessor;
+  public function __construct(CertificateLocatorHelper $certificateLocator, Settings $settings) {
     $this->certificateLocator = $certificateLocator;
     $this->settings = $settings;
   }
@@ -55,7 +51,6 @@ class OrganisationHelper {
   // phpcs:ignore
   private function getSF1500(): SF1500 {
     if (NULL === $this->sf1500) {
-
       $soapClient = new SoapClient([
         'cache_expiration_time' => explode(PHP_EOL, $this->settings->getCacheExpiration()),
       ]);
@@ -69,11 +64,9 @@ class OrganisationHelper {
 
       $sf1514 = new SF1514($soapClient, $options);
 
-      $sf1500XMLBuilder = new SF1500XMLBuilder();
-
       unset($options['sts_applies_to']);
 
-      $this->sf1500 = new SF1500($soapClient, $sf1514, $sf1500XMLBuilder, $this->propertyAccessor, $options);
+      $this->sf1500 = new SF1500($sf1514, $options);
     }
 
     return $this->sf1500;
@@ -172,6 +165,159 @@ class OrganisationHelper {
         : $this->settings->getOrganisationProductionManagerRoleId();
 
     return $this->getSF1500()->getManagerBrugerAndFunktionsIdFromUserId($userId, $managerFunktionsTypeId);
+  }
+
+  /**
+   * Search for adresser.
+   *
+   * @param array $query
+   *   The search query.
+   *
+   * @return \ItkDev\Serviceplatformen\Service\SF1500\Model\Adresse[]
+   *   The list of results.
+   *
+   * @phpstan-param array<string, mixed> $query
+   * @phpstan-return array<string, mixed>
+   */
+  public function searchAdresse(array $query): array {
+    /** @var \ItkDev\Serviceplatformen\Service\SF1500\AdresseService $service */
+    $service = $this->getSF1500()->getService(AdresseService::class);
+    $result = $service->soeg($query);
+
+    return $result;
+  }
+
+  /**
+   * Search for bruger.
+   *
+   * @param array $query
+   *   The search query.
+   *
+   * @return \Digitaliseringskataloget\SF1500\Organisation6\Person\Person[]
+   *   The list of results.
+   *
+   * @phpstan-param array<string, mixed> $query
+   * @phpstan-return array<string, mixed>
+   */
+  public function searchBruger(array $query): array {
+    /** @var \ItkDev\Serviceplatformen\Service\SF1500\BrugerService $service */
+    $service = $this->getSF1500()->getService(BrugerService::class);
+    return $service->soeg($query, [
+      Bruger::FIELD_EMAIL,
+      Bruger::FIELD_MOBILTELEFON,
+      Bruger::FIELD_LOKATION,
+    ]);
+  }
+
+  /**
+   * Search for persons.
+   *
+   * @param array $query
+   *   The search query.
+   *
+   * @return \Digitaliseringskataloget\SF1500\Organisation6\Person\Person[]
+   *   The list of results.
+   *
+   * @phpstan-param array<string, mixed> $query
+   * @phpstan-return array<string, mixed>
+   */
+  public function searchPerson(array $query): array {
+    /** @var \ItkDev\Serviceplatformen\Service\SF1500\PersonService $service */
+    $service = $this->getSF1500()->getService(PersonService::class);
+    $result = $service->soeg($query);
+
+    return $result;
+  }
+
+  /**
+   * Search for brugere.
+   *
+   * @param array $query
+   *   The query.
+   *
+   * @return \ItkDev\Serviceplatformen\Service\SF1500\Model\AbstractModel[]|array
+   *   The search result.
+   *
+   * @phpstan-param array<string, mixed> $query
+   */
+  public function search(array $query): array {
+    $query += [
+      AbstractService::PARAMETER_LIMIT => 10,
+    ];
+
+    $results = [];
+
+    $adresseQuery = $this->getSubQuery(AdresseService::getValidFilters(), $query);
+    if (!empty($adresseQuery)) {
+      // @todo Handle address search.
+    }
+
+    $brugerQuery = $this->getSubQuery(BrugerService::getValidFilters(), $query);
+    if (!empty($brugerQuery)) {
+      $results[] = $this->searchBruger($brugerQuery);
+    }
+
+    $personQuery = $this->getSubQuery(PersonService::getValidFilters(), $query);
+    if (!empty($personQuery)) {
+      $personResult = $this->searchPerson($personQuery);
+      if (!empty($personResult)) {
+        // phpcs:disable
+        // Filtering on multiple person ids in one go does not seem to work
+        // (only a single result is returned), so we make a lot of calls …
+        // $results[] = $this->searchBruger([
+        //   BrugerService::FILTER_PERSON_ID => array_map(
+        //     static fn(Person $person) => $person->id,
+        //     $personResult
+        //   )
+        // ]);
+        // phpcs:enable
+        foreach ($personResult as $person) {
+          $results[] = $this->searchBruger([
+            BrugerService::FILTER_PERSON_ID => $person->id,
+          ]);
+        }
+      }
+    }
+
+    // Index by id.
+    $result = [];
+    foreach (array_merge(...$results) as $model) {
+      $result[$model->id] = $model;
+    }
+
+    return array_values($result);
+  }
+
+  /**
+   * Get a sub-query containing only the specified (valid) filters.
+   *
+   * @param array $validFilters
+   *   The valid filters.
+   * @param array $query
+   *   The full query.
+   *
+   * @phpstan-param array<string> $validFilters
+   * @phpstan-param array<string, mixed> $query
+   * @phpstan-return array<string, mixed>
+   */
+  private function getSubQuery(array $validFilters, array &$query): array {
+    $subQuery = [];
+    foreach ($validFilters as $filter) {
+      if (isset($query[$filter])) {
+        $subQuery[$filter] = $query[$filter];
+        unset($query[$filter]);
+      }
+    }
+
+    if (!empty($subQuery)) {
+      foreach (AbstractService::getPaginationParameters() as $parameter) {
+        if (isset($query[$parameter])) {
+          $subQuery[$parameter] = $query[$parameter];
+        }
+      }
+    }
+
+    return $subQuery;
   }
 
 }